refactor(exec-rs): apply code-review touch-ups (no cast, no direct reqwest)
- Remove `as i64` cast on bot_guid in nav.find_path json! call; u64 is
  passed directly, matching lfg-matchmaker harness convention. The test
  assertion `assert_eq!(body["bot_guid"], 42i64)` remains green because
  serde_json Value::Number equality is numeric, not typed.
- Drop direct `reqwest` dependency from exec-rs Cargo.toml — all HTTP is
  owned by tot-harness-client. Add comment explaining the tokio/tracing
  deps are staged for the future graceful-shutdown binary, not M0 lib.

8/8 tests green, clippy clean, build clean.

+++ tot/exec-rs/src/nav.rs
@@ -71,7 +71,7 @@ pub async fn find_path(
         .call(
             "nav.find_path",
             json!({
-                "bot_guid": bot_guid as i64,
+                "bot_guid": bot_guid,
                 "dest_x": dest.x,
                 "dest_y": dest.y,
                 "dest_z": dest.z,
